integrate auth feature into new DataStore
